third_party/blink/web_tests/external/wpt/html/rendering/widgets/appearance/default-styles-expected.txt

This is a testharness.js-based test.
FAIL <input> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="text"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="TEXT"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="search"> assert_equals: -webkit-appearance expected "auto" but got "searchfield"
FAIL <input type="tel"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="url"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="email"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="password"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="date"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="month"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="week"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="time"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="datetime-local"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="number"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="range"> assert_equals: -webkit-appearance expected "auto" but got "slider-horizontal"
FAIL <input type="color"> assert_equals: -webkit-appearance expected "auto" but got "square-button"
FAIL <input type="checkbox"> assert_equals: -webkit-appearance expected "auto" but got "checkbox"
FAIL <input type="radio"> assert_equals: -webkit-appearance expected "auto" but got "radio"
FAIL <input type="submit"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="reset"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="button"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="unknowntype"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <select> assert_equals: -webkit-appearance expected "auto" but got "menulist"
FAIL <select multiple=""> assert_equals: -webkit-appearance expected "auto" but got "listbox"
FAIL <select size="2"> assert_equals: -webkit-appearance expected "auto" but got "listbox"
FAIL <button> assert_equals: -webkit-appearance expected "auto" but got "button"
FAIL <textarea> assert_equals: -webkit-appearance expected "auto" but got "textarea"
FAIL <meter> assert_equals: -webkit-appearance expected "auto" but got "meter"
FAIL <progress> assert_equals: -webkit-appearance expected "auto" but got "progress-bar"
FAIL <input type="hidden"> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <input type="HIDDEN"> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <input type="file"> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <input type="image"> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <div> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <details> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <summary> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <video> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <video controls=""> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <menuitem> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <marquee> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <keygen> assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <input> (namespace: null) assert_equals: appearance (no prefix) expected "none" but got ""
FAIL <input> (namespace: http://www.w3.org/2000/svg) assert_equals: appearance (no prefix) expected "none" but got ""
FAIL The html element assert_equals: appearance (no prefix) expected "none" but got ""
FAIL The body element assert_equals: appearance (no prefix) expected "none" but got ""
Harness: the test ran to completion.
